# specs/openapi.yaml
openapi: 3.0.3
info:
  title: HouseFlow API
  description: |
    API pour l'application HouseFlow - Suivi d'entretien de maison.

    ## MVP Features
    - Authentification (inscription, connexion, déconnexion)
    - Gestion des maisons (CRUD)
    - Gestion des appareils (CRUD)
    - Gestion des types d'entretien (CRUD)
    - Logging des entretiens effectués
    - Calcul des scores de progression

    ## Conventions
    - **Erreurs** : ProblemDetails (`application/problem+json`, schéma `ErrorResponse`) avec un
      `code` machine quand l'interface doit réagir (voir `ErrorResponse.code`).
    - **403 / 404** : 404 (`not_found`) si la ressource — ou la maison qui la contient — n'existe
      pas ; 403 (`forbidden`) si elle existe mais que l'appelant n'a pas le rôle requis (R5).
    - **Dates calendaires** (date d'entretien, échéance, date d'installation) : un jour du
      calendrier Europe/Paris, transporté comme **minuit UTC de ce jour**
      (`yyyy-MM-ddT00:00:00Z`). Les réponses l'envoient toujours sous cette forme
      (`format: date-time`) ; les requêtes acceptent `yyyy-MM-dd` ou `yyyy-MM-ddT00:00:00Z`
      (`format: date`) — seule la partie date est prise en compte, jamais un instant local.
  version: 1.0.0
  contact:
    name: HouseFlow Team
servers:
  - url: http://localhost:5203/api/v1
    description: Local Development Server
  - url: https://api.houseflow.app/v1
    description: Production Server
...
paths:
...
  /maintenance-types/{typeId}:
    parameters:
      - $ref: '#/components/parameters/typeId'

    put:
      operationId: updateMaintenanceType
      summary: Modifier un type d'entretien
      description: |
        Mise à jour partielle. Changer la périodicité recalcule l'échéance depuis le dernier
        enregistrement (l'échéance est toujours dérivée, jamais stockée). Passer à une
        périodicité non `Custom` efface `customDays` / `customMonths`.
      tags: [MaintenanceTypes]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMaintenanceTypeRequest'
      responses:
        '200':
          description: Type modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceType'
        '400':
          description: Données invalides (passage à Custom sans intervalle…)
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteMaintenanceType
      summary: Supprimer un type d'entretien
      description: Supprime un type et son historique (cascade).
      tags: [MaintenanceTypes]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Type supprimé
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # MAINTENANCE INSTANCES (US-033 (#179), US-034 (#180))
  # ============================================================================
  /maintenance-types/{typeId}/instances:
    parameters:
      - $ref: '#/components/parameters/typeId'

    post:
      operationId: logMaintenance
      summary: Logger un entretien
      description: |
        Enregistre qu'un entretien a été effectué (propriétaire, collaborateur RW, locataire).
        Met automatiquement à jour le statut du type et calcule la prochaine échéance (R2 :
        date de réalisation + périodicité). La date ne peut pas être postérieure à la date du
        jour à Europe/Paris.
      tags: [MaintenanceInstances]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogMaintenanceRequest'
            example:
              date: "2024-10-12"
              cost: 95.00
              provider: "Ramoneur Express"
              notes: "RAS, conduit en bon état"
      responses:
        '201':
          description: Entretien loggé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceInstance'
        '400':
          description: Données invalides (date future…)
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /maintenance-instances/{instanceId}:
    parameters:
      - name: instanceId
        in: path
        required: true
        description: ID de l'enregistrement d'entretien (UUID)
        schema:
          type: string
          format: uuid

    put:
      operationId: updateMaintenanceInstance
      summary: Modifier un enregistrement d'entretien
      description: |
        Modification (M3) : propriétaire, collaborateur RW ou
        locataire (R5). L'échéance du type est recalculée (R2).
      tags: [MaintenanceInstances]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMaintenanceInstanceRequest'
      responses:
        '200':
          description: Enregistrement modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceInstance'
        '400':
          description: Données invalides (date future…)
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteMaintenanceInstance
      summary: Supprimer un enregistrement d'entretien
      description: |
        Propriétaire ou collaborateur RW (pas le locataire, R5). Supprimer le dernier
        enregistrement recalcule l'échéance depuis le précédent, ou applique la règle « sans
        historique » s'il n'y en a plus (R2). Sert à « Annuler » dans le toast C5.
      tags: [MaintenanceInstances]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Enregistrement supprimé
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # DASHBOARD / UPCOMING TASKS
  # ============================================================================
  /dashboard:
    get:
      operationId: getDashboard
      summary: Accueil (P07) — entretiens à traiter et compteurs
      description: |
        Sur toutes les maisons visibles par l'utilisateur :
        - `tasks` : **tous** les entretiens « À traiter » (en retard + à faire dans les 30 jours,
          R1), sans limite, triés par échéance croissante (les retards d'abord) ;
        - compteurs R3 : `toHandleCount` (= badge de navigation), `overdueCount`,
          `pendingCount`, `upToDateCount`, `totalCount` (« {à jour}/{total} à jour ») ;
        - `nextTask` : le prochain entretien à jour (échéance la plus proche), pour l'état
          « Tout est à jour · Prochain : … » ; null s'il n'y en a aucun.
      tags: [MaintenanceTypes]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Données de l'accueil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
  /upcoming-tasks:
    get:
      operationId: getUpcomingTasks
      summary: Liste des tâches d'entretien à venir
      description: |
        Retourne toutes les tâches d'entretien à faire (30 jours) ou en retard
        pour l'ensemble des maisons et appareils de l'utilisateur, triées par échéance
        croissante (les retards d'abord). Préférer `GET /dashboard` pour l'accueil.
      tags: [MaintenanceTypes]
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Nombre maximum de tâches à retourner
          required: false
          schema:
            type: integer
            minimum: 1
            example: 5
      responses:
        '200':
          description: Liste des tâches à venir
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpcomingTasksResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via /auth/login ou /auth/register
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clé API générée depuis les paramètres utilisateur
  parameters:
...
    typeId:
      name: typeId
      in: path
      required: true
      description: ID du type d'entretien (UUID)
      schema:
        type: string
        format: uuid
...
  responses:
    Forbidden:
      description: La ressource existe mais l'appelant n'a pas le rôle requis (`forbidden`, R5)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ressource inconnue, ou maison inexistante (`not_found`)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      description: |
        Erreur au format ProblemDetails (RFC 9457, `application/problem+json`). Le frontend
        branche sur `status` et `code`, jamais sur le texte (`detail` est en anglais, non
        destiné à l'affichage). Les 400 de validation de modèle ajoutent `errors`.
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
        detail:
          type: string
          nullable: true
          description: Message technique (anglais)
        code:
          type: string
          nullable: true
          description: |
            Code machine stable, quand l'interface doit réagir spécifiquement :
            `invalid_credentials`, `account_restricted`, `invalid_refresh_token`, `email_taken`,
            `export_rate_limited`, `invitation_invalid`, `invitation_email_mismatch`,
            `invitation_already_pending`, `invitation_limit_reached`, `already_member`,
            `own_invitation`, `forbidden`, `not_found`.
        errors:
          type: object
          nullable: true
          additionalProperties:
            type: array
            items:
              type: string
          description: Erreurs de validation par champ (400 de validation uniquement)

    # ==========================================================================
    # AUTH
    # ==========================================================================
...
    Capabilities:
      type: object
      description: |
        Droits de l'appelant sur la maison (règle R5). Une action non autorisée est masquée.
      required: [canLogMaintenance, canEditDevices, canDelete, canManageHouse, canManageMembers, canViewCosts]
      properties:
        canLogMaintenance:
          type: boolean
          description: « C'est fait », enregistrer / modifier un enregistrement (propriétaire, RW, locataire)
        canEditDevices:
          type: boolean
          description: Ajouter / modifier un appareil ou un type d'entretien (propriétaire, RW)
        canDelete:
          type: boolean
          description: Supprimer un appareil, un type ou un enregistrement (propriétaire, RW)
        canManageHouse:
          type: boolean
          description: Modifier / supprimer la maison (propriétaire)
        canManageMembers:
          type: boolean
          description: Gérer membres et invitations (propriétaire)
        canViewCosts:
          type: boolean
          description: Voir coûts et prestataires (tous sauf locataire sans ce droit)
...
    UpdateMaintenanceTypeRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          nullable: true
        periodicity:
          allOf:
            - $ref: '#/components/schemas/Periodicity'
          nullable: true
        customDays:
          type: integer
          minimum: 1
          maximum: 3650
          nullable: true
        customMonths:
          type: integer
          minimum: 1
          maximum: 120
          nullable: true
    Periodicity:
      type: string
      enum: [Annual, Semestrial, Quarterly, Monthly, Biennial, Custom]
      description: |
        Calcul en mois calendaires depuis la date du dernier entretien :
        - Annual : 1 an ; Semestrial : 6 mois ; Quarterly : 3 mois ; Monthly : 1 mois ;
          Biennial : 2 ans
        - Custom : `customMonths` mois, ou `customDays` jours (historique)
    MaintenanceType:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        periodicity:
          $ref: '#/components/schemas/Periodicity'
        customDays:
          type: integer
          nullable: true
        customMonths:
          type: integer
          nullable: true
        deviceId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
    LogMaintenanceRequest:
      type: object
      required: [date]
      properties:
        date:
          type: string
          format: date
          description: Date de l'entretien effectué, pas dans le futur (Europe/Paris). Jour calendaire, `yyyy-MM-dd` ou `yyyy-MM-ddT00:00:00Z` (voir Conventions)
        cost:
          type: number
          format: decimal
          minimum: 0
          nullable: true
          description: Coût de l'entretien en euros
        provider:
          type: string
          maxLength: 200
          nullable: true
          description: Nom du prestataire
        notes:
          type: string
          maxLength: 2000
          nullable: true
          description: Notes et observations
    UpdateMaintenanceInstanceRequest:
      type: object
      description: |
        Remplacement (M3 en modification) : `cost`, `provider` et `notes` prennent la valeur envoyée —
        omis ou null les efface. `date` (obligatoire sur l'enregistrement) omise ou null est conservée.
        Un appelant qui ne voit pas les coûts (locataire sans `canViewCosts`) ne peut ni lire ni
        modifier `cost` / `provider` : ces deux champs sont alors ignorés et conservés.
      properties:
        date:
          type: string
          format: date
          nullable: true
          description: Pas dans le futur (Europe/Paris). Omise ou null = conservée. Jour calendaire, `yyyy-MM-dd` ou `yyyy-MM-ddT00:00:00Z` (voir Conventions)
        cost:
          type: number
          format: decimal
          minimum: 0
          nullable: true
        provider:
          type: string
          maxLength: 200
          nullable: true
        notes:
          type: string
          maxLength: 2000
          nullable: true
    MaintenanceInstance:
      type: object
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date-time
          description: Jour calendaire, minuit UTC (`yyyy-MM-ddT00:00:00Z`, voir Conventions)
        cost:
          type: number
          format: decimal
          nullable: true
        provider:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        maintenanceTypeId:
          type: string
          format: uuid
        maintenanceTypeName:
          type: string
          description: Nom du type (pour affichage dans historique)
        createdAt:
          type: string
          format: date-time
...
    UpcomingTask:
      type: object
      properties:
        maintenanceTypeId:
          type: string
          format: uuid
        maintenanceTypeName:
          type: string
        deviceId:
          type: string
          format: uuid
        deviceName:
          type: string
        deviceType:
          type: string
        houseId:
          type: string
          format: uuid
        houseName:
          type: string
        status:
          type: string
          enum: [pending, overdue, up_to_date]
          description: up_to_date seulement pour `Dashboard.nextTask`
        nextDueDate:
          type: string
          format: date-time
          description: Jamais nulle (R2). Jour calendaire, minuit UTC (`yyyy-MM-ddT00:00:00Z`, voir Conventions)
        lastMaintenanceDate:
          type: string
          format: date-time
          nullable: true
          description: Jour calendaire, minuit UTC (`yyyy-MM-ddT00:00:00Z`, voir Conventions)
        periodicity:
          $ref: '#/components/schemas/Periodicity'
        customDays:
          type: integer
          nullable: true
        customMonths:
          type: integer
          nullable: true
        canLogMaintenance:
          type: boolean
          description: L'appelant peut faire « C'est fait » sur cet entretien (R5)
        capabilities:
          $ref: '#/components/schemas/Capabilities'
          description: |
            Droits de l'appelant sur la maison de l'entretien (R5) — ex. `canViewCosts` pour afficher
            prestataire et coût dans M3 depuis P07.
    UpcomingTasksResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/UpcomingTask'
        overdueCount:
          type: integer
          description: Nombre de tâches en retard
        pendingCount:
          type: integer
          description: Nombre de tâches à faire (hors retards)
    Dashboard:
      type: object
      description: Accueil (P07) — voir `GET /dashboard`
      properties:
        tasks:
          type: array
          description: Tous les entretiens à traiter (retard + 30 jours), échéance croissante
          items:
            $ref: '#/components/schemas/UpcomingTask'
        toHandleCount:
          type: integer
          description: À traiter = en retard + à faire (titre de l'accueil et badge de navigation)
        overdueCount:
          type: integer
        pendingCount:
          type: integer
        upToDateCount:
          type: integer
        totalCount:
          type: integer
          description: Nombre total de types d'entretien visibles
        nextTask:
          allOf:
            - $ref: '#/components/schemas/UpcomingTask'
          nullable: true
          description: Prochain entretien à jour (échéance la plus proche), null s'il n'y en a pas

    # ==========================================================================
    # MEMBERS & INVITATIONS
    # ==========================================================================
...